@@ -2,9 +2,9 @@ namespace ArticleService.Entities;
 
 public class Article
 {
-    public Guid ArticleId { get; set; } // GUID
+    public Guid ArticleId { get; set; } 
     public string Title { get; set; } = string.Empty;
     public string Contents { get; set; } = string.Empty;
     public DateTime PublishingDate { get; set; }
-    public Guid AuthorId { get; set; } // FK
+    public Guid AuthorId { get; set; } 
 }
